Added a way to set params to the paginator.
